backend/requirements.txt
fastapi
uvicorn
pydantic[email]
python-multipart
python-dotenv
alembic
python-jose[cryptography]
passlib[bcrypt]
email-validator 
httpx
